v0.0.10

+++ src/app/app.component.ts
@@ -3,10 +3,11 @@ import { DecimalPipe } from '@angular/common';
 import { Transaction, StockPosition, TransactionType } from './shared/models/transaction';
 import { FinanceService } from './shared/services/finance.service';
 import { saveAs } from 'file-saver/FileSaver';
-import { StockService } from './shared/services/stock.service';
+import { RobinhoodService } from './shared/services/robinhood.service';
 import { IntervalObservable } from "rxjs/observable/IntervalObservable";
 import { UtilService } from './shared/services/util.service';
 import * as $ from 'jquery';
+import { AlphavantageService } from './shared/services/alphavantage.service';
 
 @Component({
   selector: 'app-root',
@@ -23,43 +24,61 @@ export class AppComponent {
   sortDir: number = 1;
   firstLoad: boolean = true;
 
-  constructor(private financeService: FinanceService, private stockService: StockService, private utilService: UtilService) {
+  constructor(private financeService: FinanceService, 
+              private stockService: RobinhoodService, 
+              private avService:AlphavantageService,
+              private utilService: UtilService) {
     this.InitPositions();
 
-    IntervalObservable.create(30000)// get our data every subsequent 10 seconds
+    IntervalObservable.create(2000)// get our data every subsequent 10 seconds
       .subscribe(() => {
         if (this.alive && (document.visibilityState != "hidden")) {
           this.getCurrentPrice();
         }
         else {
-          if (this.firstLoad) { this.getCurrentPrice(); this.firstLoad = false; }
+          if (this.firstLoad) { this.getCurrentPrice();this.firstLoad = false; }
         }
       });
+      /* IntervalObservable.create(100000)// get our data every subsequent 10 seconds
+      .subscribe(() => {this.getSMEData();}); */
   }
   private InitPositions() {
     this.stockService.GetNYSEStatus().subscribe(d => {
-      this.alive = d.is_open && (new Date(d.closes_at).valueOf() > new Date().valueOf());
+      this.alive = 
+      d.is_open && ( (new Date(d.opens_at).valueOf() < new Date().valueOf()) 
+                     && (new Date(d.closes_at).valueOf() > new Date().valueOf())
+                  );
       (this.alive) ? console.log('Open') : console.log('Closed');
     })
     this.positions = this.financeService.getAllPositions();
     this.getCurrentPrice();
   }
-  handleAddTrans(newTrans:Transaction){
+  handleAddTrans(newTrans: Transaction) {
     this.positions = this.financeService.getAllPositions();
     this.firstLoad = true;
+  }
+  getSMEData(){
+    this.positions.forEach((e,i) => 
+    {
+      setTimeout(()=>{ 
+        if (this.alive && (document.visibilityState != "hidden")) {
+          this.avService.getStockSME(e.symbol).subscribe(dat=>console.log(dat["Technical Analysis: SMA"]));
+        }
+      },5000*i);
+    });
   }
   getCurrentPrice() {
     var syms = [];
     this.positions.forEach(e => syms.push(e.symbol));
-    if(syms.length > 0){
-      this.stockService.GetTradingAPI(syms).subscribe(data => {
+    if (syms.length > 0) {
+      this.stockService.GetStockQuotes(syms).subscribe(data => {
         data.results.forEach(k => {
-          this.positions.find(e => e.symbol === k.symbol).quote = k.last_extended_hours_trade_price || k.last_trade_price;
+          this.positions.find(e => e.symbol === k.symbol).quote = k.last_trade_price;   // k.last_extended_hours_trade_price || k.last_trade_price;
           this.positions.find(e => e.symbol === k.symbol).adj_prev_close = k.adjusted_previous_close;
         });
-      }, e=>{console.log('error occured in getting quotes');});
+      }, e => { console.log('error occured in getting quotes'); });
     }
-    else{
+    else {
       console.log('no symbols');
     }
   }
@@ -100,7 +119,11 @@ export class AppComponent {
     this.sCol = sortingCol;
   }
   removeAll() {
-    this.positions = this.financeService.removeAllPositions().getAllPositions();
+    console.log('remove all fired');
+    var deleteAll = window.confirm('Are you sure to clear all?');
+    if (deleteAll) {
+      this.positions = this.financeService.removeAllPositions().getAllPositions();
+    }
   }
   openFile(event) {
     let input = event.target;
@@ -155,7 +178,7 @@ export class AppComponent {
     );
     return totSum;
   }
-  getGrandTotalGainPer(){
+  getGrandTotalGainPer() {
     var origCos = this.getGrandCostBasis();
     return ((this.getGrandTotal() - origCos) / origCos) * 100;
   }
